requirements.txt
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
idna==3.3
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
platformdirs==2.4.0
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
reticker==3.1.3
tomli==1.2.2
tweepy==4.4.0
typing-extensions==4.0.1
urllib3==1.26.7
